+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/InsertDepartment.java
@@ -3,15 +3,15 @@
  */
 package is.codion.framework.demos.employees.testing.scenarios;
 
+import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
 import is.codion.swing.framework.model.SwingEntityModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 import static is.codion.framework.domain.entity.test.EntityTestUtil.createRandomEntity;
 
 // tag::loadTest[]
-public final class InsertDepartment extends AbstractEntityPerformer<EmployeesAppModel> {
+public final class InsertDepartment implements Performer<EmployeesAppModel> {
 
   @Override
   public void perform(EmployeesAppModel application) throws Exception {