+++ demos/manual/src/main/java/is/codion/framework/demos/manual/store/test/StoreLoadTest.java
@@ -5,6 +5,7 @@ package is.codion.framework.demos.manual.store.test;
 
 import is.codion.common.model.loadtest.LoadTest;
 import is.codion.common.model.loadtest.LoadTest.Scenario;
+import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.rmi.RemoteEntityConnectionProvider;
@@ -14,10 +15,10 @@ import is.codion.framework.demos.manual.store.model.StoreApplicationModel;
 import is.codion.swing.common.model.tools.loadtest.LoadTestModel;
 import is.codion.swing.common.ui.tools.loadtest.LoadTestPanel;
 import is.codion.swing.framework.model.SwingEntityModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 import java.util.function.Function;
 
+import static is.codion.swing.framework.model.tools.loadtest.EntityLoadTestUtil.selectRandomRow;
 import static java.util.Collections.singletonList;
 
 // tag::storeLoadTest[]
@@ -38,8 +39,8 @@ public class StoreLoadTest {
     }
   }
 
-  private static class StoreScenarioPerformer extends
-          AbstractEntityPerformer<StoreApplicationModel> {
+  private static class StoreScenarioPerformer
+          implements Performer<StoreApplicationModel> {
 
     @Override
     public void perform(StoreApplicationModel application) {